use vpclist prop
