+++ src/components/crud-form-pages/RoutingTables.js
@@ -2,7 +2,7 @@ import React from "react";
 import IcseFormTemplate from "../IcseFormTemplate";
 import PropTypes from "prop-types";
 import RoutingTableForm from "../forms/RoutingTableForm";
-import { NoVpcTile } from "./Vpc";
+import { NoVpcTile } from "./NoVpcTile";
 
 export const RoutingTables = (props) => {
   return (
@@ -42,7 +42,7 @@ export const RoutingTables = (props) => {
         hideName: true,
       }}
       overrideTile={
-        props.craig.store.json.vpcs.length === 0
+        props.vpcList.length === 0
           ? NoVpcTile("Routing Tables")
           : null
       }
