+++ src/components/crud-form-pages/Vpe.js
@@ -2,7 +2,7 @@ import React from "react";
 import IcseFormTemplate from "../IcseFormTemplate";
 import PropTypes from "prop-types";
 import VpeForm from "../forms/VpeForm";
-import { NoVpcTile } from "./Vpc";
+import { NoVpcTile } from "./NoVpcTile";
 
 export const Vpe = (props) => {
   return (
@@ -38,7 +38,7 @@ export const Vpe = (props) => {
         hideName: true,
       }}
       overrideTile={
-        props.craig.store.json.vpcs.length === 0
+        props.vpcList.length === 0
           ? NoVpcTile("Virtual Private Endpoints")
           : null
       }
